WatermarkForPictures
This is a Python script that can be used to add a custom watermark to an image.  The output is saved as a new image with a custom name.  The script also allows you to specify the position of the watermark, and to make it partially transparent. You can adjust the size of the watermark, the position of the watermark and the level of transparency as per your needs.This script can be very useful if you want to add a watermark to multiple images at once. You can also use it as a function, and you can import it in your code and call it with different images and watermarks.  You can use this script to protect your images by adding a watermark, it can also be used to add a logo or a signature to your images. The script is very easy to use, you just need to specify the path of the input image and the watermark, and then you can run the script. You can also adjust the size of the watermark, the position of the watermark and the level of transparency as per your needs.  Please note that the script is using the PIL library, so you need to have it installed in your machine, you can install it by running pip install pillow You can also find the code in the main.py file, you can download the script and run it on your local machine, or you can also fork the repository and make changes as per your needs.